更新 現股當沖計算方式 更新 現股當沖 頁面

+++ Stockcomputer/app/src/main/java/com/jackpan/stockcomputer/Data/CalculateData.java
@@ -3,7 +3,7 @@ package com.jackpan.stockcomputer.Data;
 public class CalculateData {
 
     public double stockPrice;
-    public int stockTotal;
+    public long stockTotal;
     public long price;
     public int state;
 
@@ -16,7 +16,7 @@ public class CalculateData {
         this.stockPrice = stockPrice;
     }
 
-    public void setStockTotal(int stockTotal) {
+    public void setStockTotal(long stockTotal) {
         this.stockTotal = stockTotal;
     }
